Move the Google Books lookup module into the volumes lookup package
