+++ src/katubi/volumes/lookup/__init__.py
@@ -1,1 +1,2 @@
+from ._google_books import NotFound, VolumeInfo  # noqa: F401
 from ._isbn import get_or_create_volume_for_isbn  # noqa: F401
