+++ src/tests/factories/__init__.py
@@ -1,9 +1,8 @@
 from .auth import User  # noqa: F401
 from .books import Author, Book  # noqa: F401
-from .lookup import VolumeInfo  # noqa: F401
 from .reading_events import (  # noqa: F401
     FinishedReadingEvent,
     ReadingEvent,
     StartedReadingEvent,
 )
-from .volumes import Volume  # noqa: F401
+from .volumes import Volume, VolumeInfo  # noqa: F401
